Adding support for "symbols".
From a parser's perspective, a symbol is any character sequence
consisting of uppercase and lowercase letters, numerals and the
underscore, with as extra limitation that this sequence should not start
with a numeral.  From an interpreter's perspective, a symbol is only
valid of there already exists a key with a value assigned to it.

The associated semantics of an (extended) TOML expression `k = s`, where
`k` is a valid key, and `s` is a valid symbol, is that `k` is assigned
(a copy of) the value the key with the same label as `s` holds.

An example follows:
```toml
first = 1
second = first # the key "second" holds the integer value 1
first = 2      # key "second" is still 1
```